+++ src/components/ux/components/Hero.tsx
@@ -8,9 +8,9 @@ import Grid from '@mui/material/Grid2';
 import Divider from '@mui/material/Divider';
 import { styled } from '@mui/material/styles';
 import Chip from '@mui/material/Chip';
-import { useState } from 'react';
 import CircularProgress from '@mui/material/CircularProgress';
 import Alert from '@mui/material/Alert';
+import { useDocumentSummary } from './useDocumentSummary'; // Adjust path as needed
 
 const FeatureCard = styled(Box)(({ theme }) => ({
   padding: theme.spacing(3),
@@ -25,65 +25,7 @@ const FeatureCard = styled(Box)(({ theme }) => ({
 }));
 
 export default function Hero() {
-  const [file, setFile] = useState<File | null>(null);
-  const [summary, setSummary] = useState('');
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedFile = event.target.files?.[0];
-    if (selectedFile && selectedFile.type === 'application/pdf') {
-      setFile(selectedFile);
-      setError(null);
-    } else {
-      setError('Please select a valid PDF file.');
-      setFile(null);
-    }
-  };
-
-  const handleSummarize = async () => {
-    if (!file) {
-      setError('Please upload a PDF file first.');
-      return;
-    }
-
-    setLoading(true);
-    setError(null);
-
-    const formData = new FormData();
-    formData.append('file', file);
-    formData.append('src_lang', 'eng_Latn');
-    formData.append('tgt_lang', 'eng_Latn');
-    formData.append('prompt', 'Summarize the document in 3 sentences.');
-
-    try {
-      const response = await fetch(
-        'https://slabstech-dhwani-server-workshop.hf.space/v1/document_summary_v0',
-        {
-          method: 'POST',
-          headers: {
-            accept: 'application/json',
-          },
-          body: formData,
-        }
-      );
-
-      if (!response.ok) {
-        throw new Error('Failed to fetch summary');
-      }
-
-      const data = await response.json();
-      setSummary(data.summary);
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        setError('Error fetching summary: ' + err.message);
-      } else {
-        setError('Error fetching summary: An unknown error occurred');
-      }
-    } finally {
-      setLoading(false);
-    }
-  };
+  const { file, summary, loading, error, handleFileChange, handleSummarize } = useDocumentSummary();
 
   const features = [
     {
@@ -309,7 +251,7 @@ export default function Hero() {
             </Typography>
           </Stack>
 
-          {/* Videos Section (Moved Here) */}
+          {/* Videos Section */}
           <Stack
             spacing={2}
             useFlexGap
